#22 Make minimap visible on all pages, make controller smaller
