+++ frontend/src/App.js
@@ -9,22 +9,12 @@ import List from '@mui/material/List';
 import Typography from '@mui/material/Typography';
 import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
-import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
-import Grid from '@mui/material/Grid';
-import Paper from '@mui/material/Paper';
-import Link from '@mui/material/Link';
 import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import {
-  BrowserRouter as Router,
-  Route
-} from 'react-router-dom';
 import './App.css';
 import Layout from './components/Layout';
 import Routes from './Routes';
-import HomeScreen from './screens/HomeScreen';
-import { ListItemButton, ListItemIcon, ListItemText } from '@mui/material';
 import AppMenu from './components/AppMenu';
 
 // function App() {
@@ -155,6 +145,7 @@ function App() {
         >
           <Container maxWidth="lg" sx={{ p: 4, mt: 8, mb: 4 }}>
             <Routes />
+            <Layout />
           </Container>
         </Box>
       </Box>
